skip linux test b/c @actions/cache unusable
In the test environment the ACTIONS_ variables are not set
and trying to avoid the use of @actions/cache via the
isFeaturAvailable() API does not yield a situation where
we can simply run the test with caching disabled on the
unit.  There are many github issues out about this with
no great solutions...usually involves standing up one's
own caching server and then setting those vars manually.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>