making linux cache key relate to meson version too
these dependabot updates uncovered that the meson version
being installed changes slightly, sometimes a patch release,
which makes the cached build directory be invalid.  We either
need to lock down to a version of meson, or since it's rather
inconsequential, let it roll forward until we find a meson
version incompatibility with the gstreamer source.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>